channel controls module
