+++ src/app/modules/controls-channel/components/index.ts
@@ -1,0 +1,1 @@
+export * from './channel-settings';
