+++ src/app/modules/controls-channel/reducers/index.ts
@@ -1,0 +1,1 @@
+export * from './channel-controls.reducer';
