+++ src/app/modules/controls/actions/controls.actions.ts
@@ -17,8 +17,6 @@ export const searchGraph = createAction(
     props<{ searchText: string }>(),
 );
 
-export const renderEdges = createAction('[controls] renderEdges', props<{ value: boolean }>());
-
 export const renderLabels = createAction('[controls] renderLabels', props<{ value: boolean }>());
 
 export const sortOrderChange = createAction(
@@ -26,34 +24,9 @@ export const sortOrderChange = createAction(
     props<{ ascending: boolean }>(),
 );
 
-export const minEdgesRecompute = createAction(
-    '[controls] setMinimumEdges',
-    props<{ minEdges: number }>(),
-);
-
 export const setCameraFov = createAction('[controls] setCameraFov', props<{ value: number }>());
 
 export const setCameraFocusMode = createAction(
     '[controls] setCameraFocusMode',
     props<{ value: CameraFocusMode }>(),
-);
-
-export const setEdgeUseDottedLine = createAction(
-    '[controls] setEdgeUseDottedLine',
-    props<{ value: boolean }>(),
-);
-
-export const setEdgeUseDepthTest = createAction(
-    '[controls] setEdgeUseDepthTest',
-    props<{ value: boolean }>(),
-);
-
-export const capacityFilterEnable = createAction(
-    '[controls] capacityFilterEnable',
-    props<{ value: boolean }>(),
-);
-
-export const capacityFilterAmount = createAction(
-    '[controls] capacityFilterAmount',
-    props<{ value: number }>(),
 );
