+++ src/app/modules/controls/components/quick-controls/quick-controls.component.ts
@@ -3,16 +3,9 @@ import { MatCheckboxChange } from '@angular/material/checkbox';
 import { MatSlideToggleChange } from '@angular/material/slide-toggle';
 import { MatSliderChange } from '@angular/material/slider';
 import { Store } from '@ngrx/store';
-import {
-    capacityFilterAmount,
-    capacityFilterEnable,
-    minEdgesRecompute,
-    renderEdges,
-    searchGraph,
-    sortOrderChange,
-} from '../../actions';
-import * as controlSelectors from '../../selectors/controls.selectors';
-import * as controlActions from '../../actions/controls.actions';
+import { searchGraph, sortOrderChange } from '../../actions';
+import * as channelControlSelectors from '../../../controls-channel/selectors';
+import * as channelControlActions from '../../../controls-channel/actions';
 import { ScreenSizeService } from 'src/app/services/screen-size.service';
 import { GraphState } from 'src/app/modules/graph-renderer/reducer';
 import {
@@ -21,7 +14,8 @@ import {
 } from 'src/app/modules/graph-renderer/selectors';
 import { renderNodes, gotoNode } from 'src/app/modules/controls-node/actions';
 import { shouldRenderNodes } from 'src/app/modules/controls-node/selectors/node-controls.selectors';
-import { shouldRenderEdges } from '../../selectors/controls.selectors';
+import { shouldRenderEdges } from 'src/app/modules/controls-channel/selectors';
+import { minEdgesRecompute, renderEdges } from 'src/app/modules/controls-channel/actions';
 
 @Component({
     selector: 'app-quick-controls',
@@ -36,8 +30,8 @@ export class QuickControlsComponent {
     public nodeSearchResults$ = this.store$.select(selectNodesSearchResults);
     public shouldRenderEdges$ = this.store$.select(shouldRenderEdges);
     public shouldRenderNodes$ = this.store$.select(shouldRenderNodes);
-    public shouldFilterCapacity$ = this.store$.select(controlSelectors.capacityFilterEnable);
-    public capacityFilterAmount$ = this.store$.select(controlSelectors.capacityFilterAmount);
+    public shouldFilterCapacity$ = this.store$.select(channelControlSelectors.capacityFilterEnable);
+    public capacityFilterAmount$ = this.store$.select(channelControlSelectors.capacityFilterAmount);
 
     public selectFinalMatcheNodesFromSearch$ = this.store$.select(selectFinalMatcheNodesFromSearch);
 
@@ -65,11 +59,13 @@ export class QuickControlsComponent {
     }
 
     updateFilterCapacityEnable(event: MatCheckboxChange) {
-        this.store$.dispatch(controlActions.capacityFilterEnable({ value: event.checked }));
+        this.store$.dispatch(channelControlActions.capacityFilterEnable({ value: event.checked }));
     }
 
     setCapacityFiliterAmount(event: MatSliderChange) {
-        this.store$.dispatch(controlActions.capacityFilterAmount({ value: event.value || 1 }));
+        this.store$.dispatch(
+            channelControlActions.capacityFilterAmount({ value: event.value || 1 }),
+        );
     }
 
     gotoNode() {
