+++ src/app/modules/controls/selectors/controls.selectors.ts
@@ -5,8 +5,6 @@ export const controlsSelector = createFeatureSelector<GenericControlsState>('gen
 
 export const selectSearchString = createSelector(controlsSelector, (state) => state.searchText);
 
-export const shouldRenderEdges = createSelector(controlsSelector, (state) => state.renderEdges);
-
 export const shouldRenderLabels = createSelector(controlsSelector, (state) => state.renderLabels);
 
 export const selectCameraFov = createSelector(controlsSelector, (state) => state.cameraFov);
@@ -14,21 +12,4 @@ export const selectCameraFov = createSelector(controlsSelector, (state) => state
 export const selectCameraFocusMode = createSelector(
     controlsSelector,
     (state) => state.cameraFocusMode,
-);
-
-export const selectEdgeDepthTest = createSelector(controlsSelector, (state) => state.edgeDepthTest);
-
-export const selectEdgeDottedLine = createSelector(
-    controlsSelector,
-    (state) => state.edgeDottedLine,
-);
-
-export const capacityFilterAmount = createSelector(
-    controlsSelector,
-    (state) => state.capacityFilterAmount,
-);
-
-export const capacityFilterEnable = createSelector(
-    controlsSelector,
-    (state) => state.capacityFilterEnable,
 );
